azure-sqldb-pyodbc/pyproject.toml
[tool.poetry]
name = "azure-sqldb-pyodbc"
version = "0.1.0"
description = ""
authors = ["Katz Sakai <sakai@onflow.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
azure-identity = "^1.5.0"
pyodbc = "^4.0.30"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.1"
mypy = "^0.812"
black = "^21.5b0"
pytest = "^6.2.4"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
